update: move gvk into a sub package under cluster
